split tests from main module files
